build: Dependency Updates + expand repo filtering (#23)
* build: Dependency Updates (contains jUnit v4 to v6)

* build: Switch from lzma:lzma to org.tukaani:xz

Implements #10.
Implements #20.

* build: Setup repository filtering to speed up resolution

* build: Update jgltf from v2 to v3

* build: Remove unnecessaly mapping block

* fix: Update junit assumeFalse to v6

+++ core/src/main/java/com/mndk/bteterrarenderer/core/tile/ogc3dtiles/GltfModelConverter.java
@@ -59,15 +59,14 @@ public class GltfModelConverter {
     }
 
     private void convertNodeModel(NodeModel nodeModel, Matrix4d parentTransform) {
-        float[] matrixArray = nodeModel.getMatrix();
+        double[] matrixArray = nodeModel.getMatrix();
         Matrix4d localTransform = new Matrix4d(parentTransform);
 
         if (matrixArray == null) {
-            // TODO: Wait for JglTF to support double precision for translation, rotation, and scale
             Matrix4d nodeTransform = new Matrix4d();
-            float[] scaleArray = nodeModel.getScale();
-            float[] rotationArray = nodeModel.getRotation();
-            float[] translationArray = nodeModel.getTranslation();
+            double[] scaleArray = nodeModel.getScale();
+            double[] rotationArray = nodeModel.getRotation();
+            double[] translationArray = nodeModel.getTranslation();
 
             if (scaleArray != null) nodeTransform.scale(new Vector3d(scaleArray));
             if (rotationArray != null) nodeTransform.rotate(JOMLUtils.quaternionXYZW(rotationArray).normalize());
